시험: #6868 시험 둘을 통합 시험으로 옮긴다 (#6868)
CI Lint 의 "Validate Rust test suite manifest" step 이 거절했다.

    [RustUnitTier] 오류: PR base 대비 source-side test 총량 증가 금지: 4207 > 4205

이 PR 이 src/parser/hwpx/section.rs 에 `#[test]` 를 둘 더했다. 이 래칫은
총량 증가도(validateTierManifestAgainstBase) 기준선 상향도 함께 막으므로
`src/` 안에서는 풀 방법이 없다 — 시험을 `tests/` 로 옮기는 것이 유일한 해다.

두 시험이 부르는 `parse_hwpx_section` 은 이미 공개 경로이고 다른 통합 시험도
같은 경로를 쓴다(tests/cases/issue_4730_hwpx_section_depth.rs). 그대로 옮겼다 —
XML 픽스처도 단언도 손대지 않았고, 들여쓰기만 모듈 밖 기준으로 폈다.

    tests/cases/issue_6868_nested_list_orphan_field_end.rs  (신규)
      issue6868_orphan_field_end_links_inside_nested_paragraph_lists
      issue6868_nested_list_does_not_borrow_outer_open_field

## 게이트

    [RustUnitTier] 확인 완료: 4205 tests  ← base 와 같다 (4207 에서 복귀)
    3 tests run: 3 passed
      옮긴 둘 + 같은 축의 기존 가드 task1556_same_paragraph_field_uses_range_not_orphan
    cargo fmt --all --check 통과 · clippy --all-targets -D warnings 통과

(cherry picked from commit 038a3b828cf2e82c4ebfb98a1cdb468476538aa0)
