+++ src/parser/hwpx/section.rs
@@ -9433,91 +9433,6 @@ mod tests {
     }
 
     #[test]
-    fn issue6868_orphan_field_end_links_inside_nested_paragraph_lists() {
-        // 다단락 누름틀이 **중첩 문단 목록 안에서** 열리고 닫힌다 (각주 subList / 표 칸).
-        // 종전에는 `link_orphan_field_ends` 가 구역 최상위 문단에만 걸려 이 종료 마커의
-        // `begin_ctrl_id` 가 0 으로 남았고, HWP5 저장기의 두 방출 지점이 모두
-        // `begin_ctrl_id != 0` 을 요구해 **끝 표시가 통째로 사라졌다**(#6868).
-        let xml = r##"<?xml version="1.0" encoding="UTF-8"?>
-<hs:sec xmlns:hp="http://www.hancom.co.kr/hwpml/2011/paragraph"
-        xmlns:hs="http://www.hancom.co.kr/hwpml/2011/section">
-  <hp:p paraPrIDRef="0" styleIDRef="0">
-    <hp:run charPrIDRef="0">
-      <hp:ctrl>
-        <hp:footNote number="1" instId="200">
-          <hp:subList>
-            <hp:p paraPrIDRef="0" styleIDRef="0"><hp:run charPrIDRef="0"><hp:ctrl><hp:fieldBegin id="555000111" type="CLICK_HERE" name="각주필드" fieldid="627272811"/></hp:ctrl><hp:t>앞</hp:t></hp:run></hp:p>
-            <hp:p paraPrIDRef="0" styleIDRef="0"><hp:run charPrIDRef="0"><hp:t>뒤</hp:t><hp:ctrl><hp:fieldEnd beginIDRef="555000111" fieldid="627272811"/></hp:ctrl></hp:run></hp:p>
-          </hp:subList>
-        </hp:footNote>
-      </hp:ctrl>
-    </hp:run>
-  </hp:p>
-</hs:sec>"##;
-        let section = parse_hwpx_section(xml).unwrap();
-        let footnote = section.paragraphs[0]
-            .controls
-            .iter()
-            .find_map(|c| match c {
-                Control::Footnote(f) => Some(f),
-                _ => None,
-            })
-            .expect("각주 컨트롤");
-        let begin_ctrl_id = match footnote.paragraphs[0].controls.first() {
-            Some(Control::Field(field)) => field.ctrl_id,
-            other => panic!("각주 첫 문단이 fieldBegin 을 갖지 않는다: {other:?}"),
-        };
-        let ofe = footnote.paragraphs[1]
-            .orphan_field_ends
-            .first()
-            .expect("각주 둘째 문단에 고아 fieldEnd 기록");
-        assert_eq!(ofe.begin_id_ref, 555_000_111);
-        assert_eq!(
-            ofe.begin_ctrl_id, begin_ctrl_id,
-            "중첩 목록 안에서도 짝 fieldBegin 의 control id 를 잇는다 (0 이면 HWP5 저장에서 끝 표시가 사라진다)"
-        );
-    }
-
-    #[test]
-    fn issue6868_nested_list_does_not_borrow_outer_open_field() {
-        // 바깥 문단이 **열어 둔** 필드를 중첩 목록의 종료 마커가 닫는 짝으로 훔치면 안 된다.
-        // 목록마다 독립적으로 이어야 한다 — 필드는 컨테이너 경계를 넘지 못한다.
-        let xml = r##"<?xml version="1.0" encoding="UTF-8"?>
-<hs:sec xmlns:hp="http://www.hancom.co.kr/hwpml/2011/paragraph"
-        xmlns:hs="http://www.hancom.co.kr/hwpml/2011/section">
-  <hp:p paraPrIDRef="0" styleIDRef="0">
-    <hp:run charPrIDRef="0"><hp:ctrl><hp:fieldBegin id="900000001" type="CLICK_HERE" name="바깥" fieldid="1"/></hp:ctrl><hp:t>바깥열림</hp:t></hp:run>
-    <hp:run charPrIDRef="0">
-      <hp:ctrl>
-        <hp:footNote number="1" instId="200">
-          <hp:subList>
-            <hp:p paraPrIDRef="0" styleIDRef="0"><hp:run charPrIDRef="0"><hp:t>안</hp:t><hp:ctrl><hp:fieldEnd beginIDRef="900000001" fieldid="1"/></hp:ctrl></hp:run></hp:p>
-          </hp:subList>
-        </hp:footNote>
-      </hp:ctrl>
-    </hp:run>
-  </hp:p>
-</hs:sec>"##;
-        let section = parse_hwpx_section(xml).unwrap();
-        let footnote = section.paragraphs[0]
-            .controls
-            .iter()
-            .find_map(|c| match c {
-                Control::Footnote(f) => Some(f),
-                _ => None,
-            })
-            .expect("각주 컨트롤");
-        let ofe = footnote.paragraphs[0]
-            .orphan_field_ends
-            .first()
-            .expect("각주 문단의 고아 fieldEnd");
-        assert_eq!(
-            ofe.begin_ctrl_id, 0,
-            "바깥 목록의 열린 필드를 중첩 목록이 짝으로 가져오지 않는다"
-        );
-    }
-
-    #[test]
     fn task1556_same_paragraph_field_uses_range_not_orphan() {
         // 동일 문단 내 begin+end 는 종전대로 field_ranges 로만 처리 (고아 0) — 회귀 가드.
         let xml = r#"<?xml version="1.0" encoding="UTF-8"?>
